Added notifications as well as due dates

+++ app.py
@@ -1,9 +1,9 @@
 from flask import Flask, render_template as rt, request, redirect, url_for, session
 from flask_sqlalchemy import SQLAlchemy
-import string
-import random
+from datetime import timedelta, date
 import hashlib
-from instance.SQLiteUtil import *
+from instance.Utils import *
+
 
 app = Flask(__name__)
 app.secret_key = "secret"
@@ -11,14 +11,14 @@ app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///db.sqlite'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db = SQLAlchemy(app)
 
-#dummy login
-users = {"admin":"AggiesR0ck$"}
 
-#To update models need to "flask shell" "from app import db" "db.create_all()"
+
 class Todo(db.Model):
     task_id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100))
     done = db.Column(db.Boolean)
+    created = db.Column(db.String)
+    dueDate = db.Column(db.String)
     userId = db.Column(db.Integer, db.ForeignKey('users.userId'), nullable=False)
     
 class Users(db.Model):
@@ -27,20 +27,18 @@ class Users(db.Model):
     password = db.Column(db.String(20))
     salt = db.Column(db.String(10))
 
-def saltShaker():
-    characters = string.ascii_letters + string.digits
-    salt = ''.join(random.choices(characters, k=5))
-    return salt
 
 @app.route('/')
 def home():
     if session:
-        todo_list=getUserTodos(getUserId(session['user'])[0])
+        todo_list=getUserTodos(session['userId'])
+        lateTodos=getLateList(todo_list)
         user = session['user']
     else:
         todo_list = None
         user = None 
-    return rt('base.html',todo_list=todo_list, user=user)
+        lateTodos = None
+    return rt('base.html',todo_list=todo_list, user=user, lateTodos=lateTodos)
 
 @app.route('/login', methods=["GET","POST"])
 def login():
@@ -50,7 +48,7 @@ def login():
 
             if validateLogin(username, password):
                 session['user'] = username
-                print(f"HERE: {username} {getUserId(username)[0]}")
+                session['userId'] = getUserId(username)[0]
                 return redirect(url_for("home"))
             else:
                 return rt("login.html", error="Invalid Login")
@@ -67,14 +65,16 @@ def register():
         if request.method == "POST":
             username = request.form.get('email')
             password = request.form.get('password')
-            #salt the password
-            salt = saltShaker()
-            hashPass = hashlib.sha256((password+salt).encode('utf-8')).hexdigest()
-            
-            new_user = Users(username=username, password=hashPass, salt=salt)
-            db.session.add(new_user)
-            db.session.commit()
-            return redirect(url_for("login"))
+            if validateEmail(username):
+                #salt the password
+                salt = saltShaker()
+                hashPass = hashlib.sha256((password+salt).encode('utf-8')).hexdigest()
+                
+                new_user = Users(username=username, password=hashPass, salt=salt)
+                db.session.add(new_user)
+                db.session.commit()
+                return redirect(url_for("login"))
+            return redirect(url_for("register"))
 
          
         return rt("register.html")    
@@ -82,7 +82,13 @@ def register():
 @app.route('/add',methods=['POST'])
 def add():
     name = request.form.get("name")
-    new_task=Todo(name=name,done=False, userId=getUserId(session['user'])[0])
+    dueIn = request.form.get("dueDate")
+    dayCreated = date.today()
+    if dueIn != None:
+        dueDate = dayCreated + timedelta(days=int(dueIn))
+    else:
+        dueDate = None
+    new_task=Todo(name=name,done=False, userId=session['userId'], created=dayCreated, dueDate=dueDate)
     db.session.add(new_task)
     db.session.commit()
     return redirect(url_for("home"))
@@ -90,16 +96,18 @@ def add():
 @app.route('/update/<int:todo_id>')
 def update(todo_id):
     todo= Todo.query.get(todo_id)
-    todo.done=not todo.done
-    db.session.commit()
+    if session['userId'] == todo.userId:
+        todo.done=not todo.done
+        db.session.commit()
     return redirect(url_for("home"))
 
 
 @app.route('/delete/<int:todo_id>')
 def delete(todo_id):
     todo= Todo.query.get(todo_id)
-    db.session.delete(todo)
-    db.session.commit()
+    if session['userId'] == todo.userId:
+        db.session.delete(todo)
+        db.session.commit()
     return redirect(url_for("home"))
